requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
Flask==3.1.2
FormEncode==2.1.1
git-filter-repo==2.47.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
maskpass==0.3.7
mysqlclient==2.2.7
PyDispatcher==2.0.7
pynput==1.8.1
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
SQLObject==3.13.0
urllib3==2.5.0
Werkzeug==3.1.3
